requirements.txt
# =============================
# FILE: requirements.txt
# =============================
# core
openai>=1.40.0
instructor>=1.5.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# pdf + utils
PyMuPDF>=1.24.0
beautifulsoup4>=4.12.0
pandas>=2.2.1
numpy>=1.26.4
scikit-learn>=1.4.0
tqdm>=4.66.0